Changing the README for the Warmup (#2)
* Update README.md

* Update README.md

* Update LICENSE

* Update LICENSE

* Update README.md

Added the Authentification and Endorsement section of the README. Still needs to be improved.

* Create PROCESS.md

* Update README.md

Add the Figma link in the README file.

* Update PROCESS.md

Corrected the typos in PROCESS

* Initialize project with provided Android example repository

Set up the base Android project using the official example repository provided
for the course. This commit establishes the foundation for our app development,
including the default structure, configuration files, and .gitignore.

From this point forward, all development will build on top of this clean
baseline.