local report repository (#70)
* feat(Repository): add ReportRepository interface

Defines a new interface to abstract Repository implementation details.
This will be useful to test without going to the database.

* feat(Repository): add ReportRepositoryFirestore

Add a new firestore repository interface.
This can be used to interact with the report database in Firestore.

* feat(Repository): add ReportRepositoryProvider

Add a way to access the Report repository.
This makes sure that the app access only on repository.

* feat(Repository): add ReportRepositoryLocal

Add a Local Repository that doesn't access FireStore
This can be used to test UI without setting values in Firestore.

* test(Repository): add test for local repo

Add test for the local repository

* style(test): remove a useless comment.

* Fix(test): Change the OverViewVM

Change the OverviewViewModel.kt so that the value can
change with updates in the repository.

* Fix(test): fix the ReportsViewModel

Change the ReportViewModel.kt so that the value can
change with updates in the repository.
Change the way the screen subscribe to value,
so that the screen can update with data change

* Fix(nav): fix Screen for ReportView

Change ReportView Screen implementation from object
to data class to allow multiple diferent ReportView.
Returns a Screen to be used with navigateTo.

* fix(repo): change the repo provided

Change the ReportRepositoryProvider.kt to give the
local repo because the firestore version is
untested.

* style(import): remove unused imports.

* del(repo): remove untested file

removes untested file from the branch so
that the branch can be merged.

* doc(repo): add doc for local repo

* doc(ReportVM): add doc for reportView ui state.

* Fix(test): apply review to Test

Modify the libraries used, the style of variable
names and the Exception thrown.

* Fix(test): readd param of screens.

add param to screens so that the test can compile.

* style: change names to be more accurate.

---------

Co-authored-by: Kevin Morgado <kevin.morgado@gmail.com>