Update CI to integrate tests  (#66)
This PR focuses on updating the build.yml file defining the github action ran on any pull requests, the CI now runs all of our 
tests and provides the line coverage to our sonarcloud service. ensuring code quality and reliability.

Their are 2 reasons why this PR changes many files other than the build.yml. First for the CI to run the tests, all are required 
to be in the correct format, the reformatting causes most of the line replacements. Second, I had to modify a few test suites
for the tests to pass while still ensuring the quality of our implementations and I had to resolve an issue on the overview 
screen that would prevent us to test most of our implementations.

This CI is based on the bootcamp CI adapted for our needs using generative AI