fix(connection): Redesign ConnectionRepository functions, add Firestore rules
Removes the "vetId" and "farmerId" from "generateCode()" and "claimCode()" respectively. This makes it clearer to the user that the function should be ran by the user, and clears the misunderstanding that a user can claim a code for another user, which is how the tests were constructed up until now. The tests now properly switch accounts between actions, to reflect how the feature is meant to be used in a real situation.