Homogenization of the code using theme. (#173)
* add the first theme implementation.

This Fixes the the themes and sets a couple starting values.
Fix the package reference.
Changed the name of the sample theme.
Set a few starting color used in some screens in Color.kt.

* feat(color): add helper function for status colors.

extend MaterialTheme to allow to have a color given a status.
Add StatusColor used to define color associated with ReportStatus.
Implement the new stausColor function to ReportViewScreen
and OverviewScreen.

* Fix(signup): edit SignUpScreen to use themes.

* Fix(signIn): edit SignInScreen to use themes.

* Fix(signIn): Remove Unused Colors

* Use themes in Map and AddReportScreen.

* Comment out OverviewScreenPreview

* Edit Color of SurfaceVariant.

SurfaceVariant is used for cards so the latest Report
card will be more visible.

* Add Background for elements on the map screen.

* Remove the Dynamic color scheme.

* Remove Unused parameter in app theme.

---------

Co-authored-by: Kevin Morgado <kevin.morgado@gmail.com>