fix(connection): Remove disallowed get
Removes a Firestore `get` that checks if a connection exists before creating it. Allowing this through rules could allow users to scrape every connection, which we don't want. Accidentally replacing a connection is not that bad IMO, all that changes is the timestamp, which doesn't even look that useful to me.