Fix/firebase connections permissions (#160)
* fix(connection): Redesign ConnectionRepository functions, add Firestore rules

Removes the "vetId" and "farmerId" from "generateCode()" and "claimCode()" respectively. This makes it clearer to the user that the function should be ran by the user, and clears the misunderstanding that a user can claim a code for another user, which is how the tests were constructed up until now. The tests now properly switch accounts between actions, to reflect how the feature is meant to be used in a real situation.

* fix(connection): General linter cleanup, ktfmtFormat

* fix(connection): Build issue due to merge

* fix(connection): Remove disallowed get

Removes a Firestore `get` that checks if a connection exists before creating it. Allowing this through rules could allow users to scrape every connection, which we don't want. Accidentally replacing a connection is not that bad IMO, all that changes is the timestamp, which doesn't even look that useful to me.

* fix(connection): Visual line cleanup

* fix(connection): Update Firestore rules

* fix(connection): Restore Connection active -> true

for some reason this line is absolutely crucial to the functionality of the code, removing it causes an inexplicable PERMISSION_DENIED exception