First Planner Implementation (#264)
* Feat(Planner): add planner tab to the bottom bar.

Allow to access the Planning screen from Overview and Map.

* Feat(Theme): add Add onStatusColor function.

Can be used for text shown above status colors.
Meant to  fix white text on yellow, when in dark
mode text would be white on IN_PROGRESS color.

* Feat(Planner): add PlannerScreen UI.

This only adds the screen with none of the logic.
Allow to test the navigation and have feedback
before complementing screen logic.

* Test(Planner): add navigation Testfor planner.

Tests the bottom bar.
Remove Sprint1 from the file and class names.
 for clarity.

* Fix(Planner): apply the requested change.

Rename preview reports.
Rename iterable for day number and letter.
Add minimum height for daily tasks so that it can display the title.
Edit the DailyTasks logic to display information depending on task height.