Fix/emulators for physical devices (#281)
* feat(EmulatorManager): Add support for standalone devices

Tries to connect to the local host, and if not found, scans the network to find an emulator on another machine. If it still fails then you really forgot to run emulators and there's nothing I can do

* feat(EmulatorManager): Implement tests to make sure emulators work

Kinda pointless tests but it makes sure emulators are running I guess. Also tests if no emulator is found if there is no internet connection, and refactors SignInScreenTest to use the same helper function

* fix(EmulatorManager): Address self review

* fix(EmulatorManager): Cleaner test, add timeout for scanning

Includes a detailed error message in case someone is unable to run tests and doesn't know how to do network magic

* ktfmtFormat