Mock AuthRepository (#358)
* Change return type of functions in AuthRepo

Make the methods that make a user log in return the uid of the
user instead of the FirebaseUser because this is very painful to
mock and we only use the uid in the app.

* Remove google sign in test in SignInTest file

Remove this test because it's a subset of one of our e2e but
mainly because I want to mock Auth in the SignInTest and this
one is not realistic.

* Add Fake repository for authentication.

Create a fake repository that mimics firebase authentication
without any firebase dependency, should be used in SignIn and
SignUp tests.

* Add Fake repository to SignInTests.

* Add function for FakeRepo to switch connection state.

* Add mocks to SignUp Tests.

* Fix uid assignment in sign-in process

* Fix uid update in signUp

* Set current user in fake repository after sign in and sign up.

---------

Co-authored-by: adam_franco <adam.franco@epfl.ch>